[UPDATE] 안내 화면 추가, 화면 분리 작업
- 로고 추가 및 화면 안내 컴포넌트 추가
- 작업별 컴포넌트 분리 작업
